splitting the cultureaction file
