+++ Assets/Resources/Scripts/CultureScripts/DecisionMaker.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 35cd8ccf13669ae4598cbff45ce881e2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
