+++ Assets/Resources/Scripts/CultureScripts/DefaultAction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 305aa494591023a48a57598f695020c5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
