+++ Assets/Resources/Scripts/CultureScripts/Turn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6beb9048e8da60247a85481cd1a25bf2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
